+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [Unreleased]
+
+### Added
+
+- Custom actions for Workflow plugins (CMEM-5576).
+
 ## [4.8.0] 2024-09-12 - shipped with DI v24.3.0
 
 ### Added